2672:check step 3 attendance details (#1043)
* Check Step 3 of Attendance details and save details successfully.

* update test.

+++ src/main/java/shamu/company/attendance/dto/EmployeeOvertimeDetailsDto.java
@@ -16,5 +16,5 @@ public class EmployeeOvertimeDetailsDto {
 
     private String overtimeLaw;
 
-    private Float regularPay;
+    private Double regularPay;
 }
